+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 
 - Swap out "loading" placeholder text with placeholder for map
+- Display pointer cursor when mouse is over collapsible sections
 
 ## [1.11.4] - 2025-10-28